actualizar

+++ src/controllers/deck.controllers.js
@@ -152,6 +152,11 @@ const addCardToDeck = async (req, res) => {
 const deleteDeck = async (req, res) => {
     try {
         const { id } = req.params;
+        const userId = req.user._id;
+        const user = await User.findById(userId);
+        if (!req.user || !req.user._id) {
+            return res.status(401).json({ error: "No autorizado" });
+        }
         const deletedDeck = await Deck.findByIdAndDelete(id);
         if (!deletedDeck) {
             return res.status(404).json({ error: "Mazo no encontrado" });
